refacto structure
