--- src/utils/components-list.ts
@@ -1,146 +1,0 @@
-export const components = [
-  "app",
-  "accordion",
-  "alert",
-  "authForm",
-  "avatar",
-  "avatarGroup",
-  "badge",
-  "banner",
-  "blogPost",
-  "blogPosts",
-  "breadcrumb",
-  "button",
-  "buttonGroup",
-  "calendar",
-  "card",
-  "carousel",
-  "changelogVersion",
-  "changelogVersions",
-  "chatMessage",
-  "chatMessages",
-  "chatPalette",
-  "chatPrompt",
-  "chatPromptSubmit",
-  "checkbox",
-  "chip",
-  "collapsible",
-  "colorModeAvatar",
-  "colorModeButton",
-  "colorModeImage",
-  "colorModeSelect",
-  "colorModeSwitch",
-  "colorPicker",
-  "commandPalette",
-  "container",
-  "contentNavigation",
-  "contentSearch",
-  "contentSearchButton",
-  "contentSurround",
-  "contentToc",
-  "contextMenu",
-  "dashboardGroup",
-  "dashboardNavbar",
-  "dashboardPanel",
-  "dashboardResizeHandle",
-  "dashboardSearch",
-  "dashboardSearchButton",
-  "dashboardSidebar",
-  "dashboardSidebarCollapse",
-  "dashboardSidebarToggle",
-  "dashboardToolbar",
-  "drawer",
-  "dropdownMenu",
-  "error",
-  "fileUpload",
-  "footer",
-  "footerColumns",
-  "form",
-  "formField",
-  "header",
-  "icon",
-  "input",
-  "inputMenu",
-  "inputNumber",
-  "inputTags",
-  "kbd",
-  "link",
-  "localeSelect",
-  "main",
-  "marquee",
-  "modal",
-  "navigationMenu",
-  "page",
-  "pageAnchors",
-  "pageAside",
-  "pageBody",
-  "pageCard",
-  "pageColumns",
-  "pageCta",
-  "pageFeature",
-  "pageGrid",
-  "pageHeader",
-  "pageHero",
-  "pageLinks",
-  "pageList",
-  "pageLogos",
-  "pageSection",
-  "pagination",
-  "pinInput",
-  "popover",
-  "pricingPlan",
-  "pricingPlans",
-  "pricingTable",
-  "progress",
-  "radioGroup",
-  "select",
-  "selectMenu",
-  "separator",
-  "skeleton",
-  "slideover",
-  "slider",
-  "stepper",
-  "switch",
-  "table",
-  "tabs",
-  "textarea",
-  "timeline",
-  "toast",
-  "tooltip",
-  "tree",
-  "user",
-];
-
-export const proseComponents = [
-  "h1",
-  "h2",
-  "h3",
-  "h4",
-  "p",
-  "a",
-  "blockquote",
-  "strong",
-  "em",
-  "ul",
-  "ol",
-  "hr",
-  "table",
-  "img",
-  "code",
-  "pre",
-  "accordion",
-  "badge",
-  "card",
-  "cardGroup",
-  "codeCollapse",
-  "codeGroup",
-  "codePreview",
-  "codeTree",
-  "collapsible",
-  "field",
-  "fieldGroup",
-  "icon",
-  "kbd",
-  "tabs",
-  "steps",
-];
